+++ dp_mobility_report/report/html/templates.py
@@ -37,6 +37,7 @@ def render_html(
     args: dict = {}
 
     args["title"] = "DP Mobility Report"
+    args["subtitle"] = dpmreport.subtitle
     args["output_filename"] = output_filename
 
     with tqdm(  # progress bar
@@ -44,6 +45,9 @@ def render_html(
     ) as pbar:
 
         args["config_segment"] = config_templates.render_config(dpmreport)
+
+        if dpmreport.privacy_budget is not None:
+            args["dp_info"] = config_templates.render_dp_info()
 
         if not set(const.OVERVIEW_ELEMENTS).issubset(dpmreport.analysis_exclusion):
             args["overview_segment"] = overview_templates.render_overview(dpmreport)
@@ -91,6 +95,7 @@ def render_benchmark_html(
     args: dict = {}
 
     args["title"] = "DP Mobility Report: Benchmark"
+    args["subtitle"] = benchmarkreport.report_base.subtitle
     args["output_filename"] = output_filename
 
     with tqdm(  # progress bar
@@ -100,6 +105,8 @@ def render_benchmark_html(
         args["config_segment"] = config_templates.render_benchmark_config(
             benchmarkreport
         )
+        args["similarity_info"] = config_templates.render_similarity_info()
+        args["dp_info"] = config_templates.render_dp_info()
 
         if not set(const.OVERVIEW_ELEMENTS).issubset(
             benchmarkreport.analysis_exclusion